requirements.txt
asgiref
certifi
charset-normalizer
cloudinary
dj-database-url
Django
django-cloudinary-storage
django-cors-headers
django-filter
django-phonenumber-field
djangorestframework
djangorestframework_simplejwt
drf-yasg
gunicorn
idna
inflection
packaging
pillow
phonenumbers
psycopg2-binary
PyJWT
python-dotenv
pytz
PyYAML
requests
six
sqlparse
tzdata
uritemplate
urllib3
whitenoise